+++ hudi-utilities/src/main/java/org/apache/hudi/utilities/streamer/BaseErrorTableWriter.java
@@ -7,16 +7,17 @@
  * "License"); you may not use this file except in compliance
  * with the License.  You may obtain a copy of the License at
  *
- *      http://www.apache.org/licenses/LICENSE-2.0
+ *   http://www.apache.org/licenses/LICENSE-2.0
  *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
  */
 
-package org.apache.hudi.utilities.deltastreamer;
+package org.apache.hudi.utilities.streamer;
 
 import org.apache.hudi.client.common.HoodieSparkEngineContext;
 import org.apache.hudi.common.config.TypedProperties;
@@ -43,8 +44,8 @@ public abstract class BaseErrorTableWriter<T extends ErrorEvent> {
   // is set to this column in case of an error
   public static String ERROR_TABLE_CURRUPT_RECORD_COL_NAME = "_corrupt_record";
 
-  public BaseErrorTableWriter(HoodieDeltaStreamer.Config cfg, SparkSession sparkSession,
-                              TypedProperties props, HoodieSparkEngineContext hoodieSparkContext, FileSystem fs) {
+  public BaseErrorTableWriter(HoodieStreamer.Config cfg, SparkSession sparkSession,
+                                   TypedProperties props, HoodieSparkEngineContext hoodieSparkContext, FileSystem fs) {
   }
 
   /**
